requirements.txt
# Core Deep Learning
torch>=2.0.0,<2.3.0
torchvision>=0.15.0,<0.18.0
torchaudio>=2.0.0,<2.3.0

# Spiking Neural Networks
spikingjelly>=0.0.0.0.14

# Chess Engine
python-chess>=1.999

# Scientific Computing
numpy>=1.24.0,<2.0.0
scipy>=1.10.0,<2.0.0

# Visualization & Monitoring
matplotlib>=3.7.0,<4.0.0
tensorboard>=2.13.0,<3.0.0
seaborn>=0.12.0

# Progress & Logging
tqdm>=4.65.0
pyyaml>=6.0

# Data Processing
pandas>=2.0.0,<3.0.0

# Testing
pytest>=7.3.0,<8.0.0

# Optional: For Jupyter notebooks (if you want to experiment)
# jupyter>=1.0.0
# ipykernel>=6.0.0
